+++ TeamProject/TeamProject/Global/Resource/Assets.xcassets/AccentColor.colorset/Contents.json
