+++ TeamProject/TeamProject/Global/Resource/Assets.xcassets/AppIcon.appiconset/Contents.json
