+++ TeamProject/TeamProject/Global/Resource/Assets.xcassets/Contents.json
